+++ src/main/kotlin/dev/aaronhowser/mods/aaron/AaronUtil.kt
@@ -1,0 +1,19 @@
+package dev.aaronhowser.mods.aaron
+
+import net.neoforged.neoforge.common.UsernameCache
+import java.util.*
+
+object AaronUtil {
+
+	fun getCachedUuid(playerUsername: String): UUID? {
+		val map = UsernameCache.getMap()
+		for ((uuid, username) in map) {
+			if (username.equals(playerUsername, ignoreCase = true)) {
+				return uuid
+			}
+		}
+
+		return null
+	}
+
+}